Move benzina stream classes to their own namespace
