+++ src/Library/Benzina/Stream/StreamInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Library\Benzina;
+namespace App\Library\Benzina\Stream;
 
 interface StreamInterface
 {
